refactor: Validator 및 EventHandler 클래스 application 패키지로 이동
